fix(messages): improve cost tracking precision and add tests
- Add unit tests for cost_calculator.py with mock-based testing
- Change cost_usd column type from FLOAT to DECIMAL(10, 6) for better
  monetary precision and to avoid floating-point rounding errors
- Add migration 006 for existing databases to convert FLOAT to DECIMAL
- Update SQLAlchemy model to use Numeric type instead of Float

+++ infra/mysql/migrations/005_add_message_metadata.sql
@@ -21,4 +21,4 @@ ALTER TABLE messages ADD COLUMN model VARCHAR(255) NULL COMMENT 'Model name used
 ALTER TABLE messages ADD COLUMN response_time_ms INT UNSIGNED NULL COMMENT 'Response time in milliseconds';
 
 -- Step 5: Add cost_usd column
-ALTER TABLE messages ADD COLUMN cost_usd FLOAT NULL COMMENT 'Cost in USD for this message';
+ALTER TABLE messages ADD COLUMN cost_usd DECIMAL(10, 6) NULL COMMENT 'Cost in USD for this message';
